+++ src/app/wallet/components/balance-card/select-duration.tsx
